[#2] HistoryInfoController의 메서드를 EmployeeController로 옮긴다
 - 요청으로 가져올 정보의 변경에 대비해 HistoryInfo 패키지를 Employee 패키지에 합친다.
 - 점진적으로 변경하기 위해 Controller부터 변경한다.
 - EmployeeController에서 두 API를 모두 처리할 수 있도록 URI 형식을
   변경했다.
 - HistoryInfoController의 의존성인 HistoryInfoService는 임시로
   EmployeeController에 추가하고 추후 HistoryInfoService 코드를 옮기고
   나면 지울 예정이다.
